requirements.txt
boto3
google-cloud-storage >= 1.7.0
requests
